requirements.txt
youtube_transcript_api
streamlit
streamlit-option-menu
google-generativeai
python-dotenv
pathlib
PyPDF2
Pillow